+++ src/components/auth/SignUp/Logo.jsx
@@ -2,14 +2,14 @@ import { useState } from 'react';
 import styled from '@emotion/styled';
 import { Button, Image } from '@mantine/core';
 
-const Logo = styled(Button)`
+const LogoButton = styled(Button)`
   box-shadow: ${({ selected }) => selected && '0px 0px 15px 10px rgba(79, 74, 216, 0.5)'};
   border-radius: 50%;
   padding: 0;
   height: 4.375rem;
 `;
 
-const LogoBtn = ({ logo: { id, name }, idx, subscribedOtt, setSubscribedOtt }) => {
+const Logo = ({ logo: { id, name }, idx, subscribedOtt, setSubscribedOtt }) => {
   const [selected, setSelected] = useState(false);
 
   const removeLogo = () => {
@@ -25,15 +25,15 @@ const LogoBtn = ({ logo: { id, name }, idx, subscribedOtt, setSubscribedOtt }) =
   };
 
   return (
-    <Logo
+    <LogoButton
       onClick={selected ? removeLogo : selectLogo}
       selected={selected}
       variant="none"
       type="button"
       disabled={idx > 5}>
       <Image width={70} id={id} src={`./assets/badges/${name}.svg`} alt="button" />
-    </Logo>
+    </LogoButton>
   );
 };
 
-export default LogoBtn;
+export default Logo;
