+++ src/components/auth/SignUp/index.js
@@ -2,6 +2,6 @@ export { default as SignupHeader } from './SignupHeader';
 export { default as SignupOption } from './SignupOption';
 export { default as SignupForm } from './SignupForm';
 
-export { default as LogoBtn } from './LogoBtn';
+export { default as Logo } from './Logo';
 export { default as Typing } from './Typing';
 export { default as SignupInput } from './SignupInput';
