+++ src/loaders/index.js
@@ -1,1 +1,2 @@
 // Suspence fallback loader 들 모아주세요
+export { default as ModalSkeleton } from './ModalSkeleton';
